requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.6
python-dotenv==1.0.1
google-generativeai==0.8.2
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.9.2
pydantic[email]==2.9.2
numpy==1.26.4
xgboost==2.1.4
scikit-learn==1.6.1
